cleanup some errors and add DestroyBy* to Droplet, change destroy to use by id

+++ api/v2/size.go
@@ -21,7 +21,7 @@ type Size struct {
 	Memory       int      `json:"memory"`
 	VCPUS        int      `json:"vcpus"`
 	Disk         int      `json:"disk"`
-	Transfer     int      `json:"transfer"`
+	Transfer     float32  `json:"transfer"`
 	PriceMonthly float32  `json:"price_monthly"`
 	PriceHourly  float32  `json:"price_hourly"`
 	Regions      []string `json:"regions"`
